database/migrations/app/1_20230805054719_new_migrate.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "profiles" ALTER COLUMN "offset" SET DEFAULT 0;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "profiles" ALTER COLUMN "offset" SET DEFAULT 1;"""
