requirements.txt
aerich==0.7.2
aiofiles==23.1.0
aiogram==3.0.0b9
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.5.0
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dictdiffer==0.9.0
environs==9.5.0
frozenlist==1.4.0
idna==3.4
iso8601==1.1.0
magic-filter==1.0.11
marshmallow==3.20.1
multidict==6.0.4
packaging==23.1
pydantic==2.1.1
pydantic-core==2.4.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tomlkit==0.12.1
tortoise-orm==0.19.3
typing-extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
yarl==1.9.2
